+++ src/binance_spot_downloader.cpp
@@ -28,14 +28,16 @@ struct BinanceSpotDownloader::P {
     std::unique_ptr<BinanceCommon> m_binanceCommon;
     mutable Semaphore m_maxConcurrentConvertJobs;
     Semaphore m_maxConcurrentDownloadJobs{3};
+    bool m_deleteDelistedData = false;
 
-    explicit P(const std::uint32_t maxJobs) : m_bnbSpotClient(std::make_unique<spot::RESTClient>("", "")),
+    explicit P(const std::uint32_t maxJobs, const bool deleteDelistedData) : m_bnbSpotClient(std::make_unique<spot::RESTClient>("", "")),
                                               m_binanceCommon(std::make_unique<BinanceCommon>(maxJobs)),
-                                              m_maxConcurrentConvertJobs(maxJobs) {
+                                              m_maxConcurrentConvertJobs(maxJobs),
+                                              m_deleteDelistedData(deleteDelistedData) {
     }
 };
 
-BinanceSpotDownloader::BinanceSpotDownloader(std::uint32_t maxJobs) : m_p(std::make_unique<P>(maxJobs)) {
+BinanceSpotDownloader::BinanceSpotDownloader(std::uint32_t maxJobs, bool deleteDelistedData) : m_p(std::make_unique<P>(maxJobs, deleteDelistedData)) {
 
 }
 
@@ -94,7 +96,7 @@ void BinanceSpotDownloader::updateMarketData(const std::string& dirPath, const s
             if (it == exchangeInfo.m_symbols.end()) {
                 symbolsToDelete.push_back(symbol);
                 spdlog::info(fmt::format(
-                    "Symbol: {} not found on Exchange, probably delisted, data files will be removed...", symbol));
+                    "Symbol: {} not found on Exchange, probably delisted", symbol));
             } else {
                 tempSymbols.push_back(it->m_symbol);
             }
@@ -183,30 +185,32 @@ void BinanceSpotDownloader::updateMarketData(const std::string& dirPath, const s
         m_p->m_binanceCommon->convertFromCSVToT6(csvFilePaths, T6Directory.string());
     }
 
-    for (const auto& symbol : symbolsToDelete) {
-        std::filesystem::path symbolFilePathCsv = finalPath;
-        std::filesystem::path symbolFilePathT6 = finalPath;
+    if (m_p->m_deleteDelistedData) {
+        for (const auto& symbol : symbolsToDelete) {
+            std::filesystem::path symbolFilePathCsv = finalPath;
+            std::filesystem::path symbolFilePathT6 = finalPath;
 
-        symbolFilePathCsv.append(CSV_SPOT_DIR);
-        symbolFilePathT6.append(T6_SPOT_DIR);
+            symbolFilePathCsv.append(CSV_SPOT_DIR);
+            symbolFilePathT6.append(T6_SPOT_DIR);
 
-        symbolFilePathCsv.append(Downloader::minutesToString(barSizeInMinutes));
-        symbolFilePathT6.append(Downloader::minutesToString(barSizeInMinutes));
+            symbolFilePathCsv.append(Downloader::minutesToString(barSizeInMinutes));
+            symbolFilePathT6.append(Downloader::minutesToString(barSizeInMinutes));
 
-        symbolFilePathCsv = symbolFilePathCsv.lexically_normal();
-        symbolFilePathT6 = symbolFilePathT6.lexically_normal();
+            symbolFilePathCsv = symbolFilePathCsv.lexically_normal();
+            symbolFilePathT6 = symbolFilePathT6.lexically_normal();
 
-        symbolFilePathCsv.append(symbol + ".csv");
-        symbolFilePathT6.append(symbol + ".t6");
+            symbolFilePathCsv.append(symbol + ".csv");
+            symbolFilePathT6.append(symbol + ".t6");
 
-        if (std::filesystem::exists(symbolFilePathCsv)) {
-            std::filesystem::remove(symbolFilePathCsv);
-            spdlog::info("Removing csv file for delisted symbol: {}, file: {}...", symbol, symbolFilePathCsv.string());
-        }
+            if (std::filesystem::exists(symbolFilePathCsv)) {
+                std::filesystem::remove(symbolFilePathCsv);
+                spdlog::info("Removing csv file for delisted symbol: {}, file: {}...", symbol, symbolFilePathCsv.string());
+            }
 
-        if (std::filesystem::exists(symbolFilePathT6)) {
-            std::filesystem::remove(symbolFilePathT6);
-            spdlog::info("Removing t6 file for delisted symbol: {}, file: {}...", symbol, symbolFilePathT6.string());
+            if (std::filesystem::exists(symbolFilePathT6)) {
+                std::filesystem::remove(symbolFilePathT6);
+                spdlog::info("Removing t6 file for delisted symbol: {}, file: {}...", symbol, symbolFilePathT6.string());
+            }
         }
     }
 }
